move demo code out of `metadatasource`
